Function : Weapon UI (3)
- 무기 레벨업 기능 구현
- 무기 레벨업 UI 연결 & 구현

+++ Assets/3_Scripts/Data/TableData/ItemTable/ItemTableClasses.cs
@@ -43,7 +43,7 @@ namespace ItemTableClassGroup
         public int attackValue;
         public int additionalEffect;
         public float additionEffectValue;
-        public float increaseAttackValue;
+        public int increaseAttackValue;
         public float increaseAdditionEffectValue;
         public int price;
         public int weaponType;
